Remove setup parameters from InstrumentDesign
